Move load_meta_section, load_yaml_file in load module.
